These tests should only run in a js build context :alien:
